update lava contract and adjust tests

+++ contracts/SafeMath.sol
@@ -2,24 +2,24 @@ pragma solidity ^0.4.11;
 
 
 library SafeMath {
-    function mul(uint256 a, uint256 b) internal returns (uint256) {
-        uint256 c = a * b;
+    function mul(uint a, uint b) internal returns (uint) {
+        uint c = a * b;
         assert(a == 0 || c / a == b);
         return c;
     }
 
-    function div(uint256 a, uint256 b) internal returns (uint256) {
-        uint256 c = a / b;
+    function div(uint a, uint b) internal returns (uint) {
+        uint c = a / b;
         return c;
     }
 
-    function sub(uint256 a, uint256 b) internal returns (uint256) {
+    function sub(uint a, uint b) internal returns (uint) {
         assert(b <= a);
         return a - b;
     }
 
-    function add(uint256 a, uint256 b) internal returns (uint256) {
-        uint256 c = a + b;
+    function add(uint a, uint b) internal returns (uint) {
+        uint c = a + b;
         assert(c >= a);
         return c;
     }
